test: Making all tests dismiss the permission alert
